wrote CardImageArray

+++ app/src/main/java/com/example/rankupandroid/rankup/Card.kt
@@ -1,6 +1,6 @@
 package com.example.rankupandroid.rankup
 
-data class Card(val value: Short) {
+data class Card(val value: Int) {
     init {
         require(value in 0..53) { "$value is out of the range of [0,54)" }
     }
